requirements_crawler.txt
crawl4ai>=0.3.0
google-cloud-storage>=2.10.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
tqdm>=4.66.0
